requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-telegram-bot==20.8
python-dotenv==1.0.1
